@@ -19,7 +19,7 @@ public class ResponseErrorHandler implements org.springframework.web.client.Resp
 
     @Override
     public boolean hasError(ClientHttpResponse response) throws IOException {
-        hadError = response.getRawStatusCode() >= 400;
+        hadError = response.getStatusCode().value() >= 400;
         return hadError;
     }
 
